chore: #58 move character detail to a folder, update button icons
