+++ resources/js/Components/Table/CharacterDetail/CharacterDetailTable.test.js
