+++ resources/js/Components/Table/CharacterDetail/CharacterDetailTable.tsx
