+++ resources/js/Components/Table/CharacterDetail/index.ts
